Implemented optional "hex" commandline paramater to display hex values on line above each instruction disassembly.
Really annoying bug when for loop finishes:

:panic: runtime error: index out of range [8192] with length 8192

If I redirect the output to a text file this doesn't happen.

If I pipe the output to less it happens sometimes, but if you pageup/pagedown execution continues as normal.

Very odd.

I'm not good enough with Golang to know what this is. Anyone reading this who knows how to fix it please let me know. Thanks!