Added a 3rd commandline paramter for hex entrypoint address
Still have that really annoying bug when the for loop finishes:

":panic: runtime error: index out of range [8192] with length 8192"

If I redirect the output to a text file this doesn't happen.

If I pipe the output to less it happens occasionally, but if you pageup/pagedown execution continues as normal.

Very odd.

I'm not good enough with Golang to know what is causing this. If anyone reading this who knows how to fix it please let me know. Thanks!